Revert "[GEOT-6383] Allow GranuleSource users to get a file oriented view, rather than a slice oriented one"
This reverts commit 053dd30e19182db32c3fdf7e318aa6275b130bed.
