+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/RasterManager.java
@@ -1561,7 +1561,7 @@ public class RasterManager implements Cloneable {
         synchronized (this) {
             if (readOnly) {
                 if (granuleSource == null) {
-                    granuleSource = new GranuleCatalogSource(this, granuleCatalog, typeName, hints);
+                    granuleSource = new GranuleCatalogSource(granuleCatalog, typeName, hints);
                     if (!typeName.equalsIgnoreCase(name)) {
                         // need to rename
                         granuleSource = new RenamingGranuleSource(name, granuleSource);
@@ -1912,23 +1912,5 @@ public class RasterManager implements Cloneable {
         }
 
         return crsAttribute;
-    }
-
-    /**
-     * The parent directory that can be used with the {@link PathType} enumeration
-     *
-     * @return
-     */
-    public String getParentLocation() {
-        return URLs.fileToUrl(getParentReader().parentDirectory).toString();
-    }
-
-    /**
-     * The attribute containing the location information for the single granules
-     *
-     * @return
-     */
-    public String getLocationAttribute() {
-        return getParentReader().locationAttributeName;
     }
 }
