+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/catalog/GranuleCatalogStore.java
@@ -45,12 +45,15 @@ public class GranuleCatalogStore extends GranuleCatalogSource implements Granule
 
     private Transaction transaction;
 
+    private RasterManager manager;
+
     public GranuleCatalogStore(
             RasterManager manager,
             GranuleCatalog catalog,
             final String typeName,
             final Hints hints) {
-        super(manager, catalog, typeName, hints);
+        super(catalog, typeName, hints);
+        this.manager = manager;
     }
 
     @Override
